period object converter utils
